requirements.txt
Django==5.2.7
black==25.9.0
django-environ==0.12.0
psycopg2-binary==2.9.11

djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2

dj-rest-auth==7.0.1
django-allauth==65.12.1
django-allauth[socialaccount]==65.12.1
djangorestframework-simplejwt==5.5.1

drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1

Pillow==12.0.0
